nlp/requirements.txt
asttokens==2.0.5
backcall==0.2.0
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cymem==2.0.6
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
executing==0.8.3
filelock==3.6.0
huggingface-hub==0.4.0
idna==3.3
ipython==8.1.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
nltk==3.7
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
preshed==3.0.6
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
stack-data==0.2.0
thinc==8.0.15
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.4.0+cpu
torchvision==0.5.0+cpu
tqdm==4.63.0
traitlets==5.1.1
transformers==4.17.0
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
wasabi==0.9.0
wcwidth==0.2.5
